Changed toOpenGLColor to use 4 bytes instead of an int to avoid endian problems. Slightly reorganized D3D9 window setup. Replaced math.h with irrMath.h in all files. Moved a helper function from CSceneCollisionManager.h to .cpp. Added some new targets in Makefile for dll compilation, though not yet useable due to lack of defines.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@184 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/COpenGLDriver.h
@@ -292,7 +292,7 @@ namespace video
 
 		core::stringw Name;
 		core::matrix4 Matrices[ETS_COUNT];
-		core::array<u32> ColorBuffer;
+		core::array<u8> ColorBuffer;
 
 		// enumeration for rendering modes such as 2d and 3d for minizing the switching of renderStates.
 		enum E_RENDER_MODE
